+++ glue_vispy_viewers/common/tests/test_vispy_viewer.py
@@ -8,7 +8,7 @@ from mock import patch
 from glue.core import Data, DataCollection
 from glue.app.qt import GlueApplication
 
-from .util import simple_session
+from glue.core.tests.util import simple_session
 
 from ..vispy_data_viewer import BaseVispyViewer
 from ...volume.volume_viewer import VispyVolumeViewer
@@ -74,6 +74,8 @@ class BaseTestDataViewer(object):
         w.state.z_max = 10.8
 
         w.state.visible_axes = False
+
+        app.close()
 
 
 class TestDataViewerVolume(BaseTestDataViewer):