+++ glue_vispy_viewers/scatter/layer_style_widget.py
@@ -4,7 +4,6 @@ import os
 
 from qtpy import QtWidgets
 
-from glue.utils import nonpartial
 from glue.utils.qt import load_ui
 from glue.external.echo.qt import autoconnect_callbacks_to_qt
 
@@ -35,10 +34,10 @@ class ScatterLayerStyleWidget(QtWidgets.QWidget):
         self._update_size_mode()
         self._update_color_mode()
 
-        self.state.add_callback('color_mode', nonpartial(self._update_color_mode))
-        self.state.add_callback('size_mode', nonpartial(self._update_size_mode))
+        self.state.add_callback('color_mode', self._update_color_mode)
+        self.state.add_callback('size_mode', self._update_size_mode)
 
-    def _update_size_mode(self):
+    def _update_size_mode(self, *args):
 
         if self.state.size_mode == "Fixed":
             self.ui.size_row_2.hide()
@@ -49,7 +48,7 @@ class ScatterLayerStyleWidget(QtWidgets.QWidget):
             self.ui.combosel_size_attribute.show()
             self.ui.size_row_2.show()
 
-    def _update_color_mode(self):
+    def _update_color_mode(self, *args):
 
         if self.state.color_mode == "Fixed":
             self.ui.color_row_2.hide()