+++ glue_vispy_viewers/volume/viewer_state.py
@@ -16,14 +16,14 @@ class Vispy3DVolumeViewerState(Vispy3DViewerState):
 
         self.update_from_dict(kwargs)
 
-    def _update_attributes(self, *args):
-
+    def _first_3d_data(self):
         for layer_state in self.layers:
             if getattr(layer_state.layer, 'ndim', None) == 3:
-                data = layer_state.layer
-                break
-        else:
-            data = None
+                return layer_state.layer
+
+    def _update_attributes(self, *args):
+
+        data = self._first_3d_data()
 
         if data is None:
 
@@ -38,3 +38,19 @@ class Vispy3DVolumeViewerState(Vispy3DViewerState):
             type(self).x_att.set_choices(self, [x_cid])
             type(self).y_att.set_choices(self, [y_cid])
             type(self).z_att.set_choices(self, [z_cid])
+
+    @property
+    def clip_limits_relative(self):
+
+        data = self._first_3d_data()
+
+        if data is None:
+            return [0., 1., 0., 1., 0., 1.]
+        else:
+            nz, ny, nx = data.shape
+            return (self.x_min / nx,
+                    self.x_max / nx,
+                    self.y_min / ny,
+                    self.y_max / ny,
+                    self.z_min / nz,
+                    self.z_max / nz)