+++ CHANGES.md
@@ -10,6 +10,17 @@
 - Improve performance for volume rendering for arrays larger than 2048
   along one or more dimensions. [#303]
 
+- Improve performance when closing a session that has large volume
+  visualizations. [#307]
+
+- Improve performance when clipping the data outside the box. [#307]
+
+- Fixed a bug that caused layers to be shown/hidden out of sync with
+  checkboxes. [#307]
+
+- Fixed a bug that caused circular references to viewers to cause issues
+  after the viewers were closed. [#307]
+
 0.9.2 (unreleased)
 ------------------